# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Junkio Expense Tracker API
  version: 1.0.0
  description: >-
    **Chào mừng đến với thư viện API của Junkio Expense Tracker!**  


    Dưới đây là các tài liệu hướng dẫn sử dụng API chi tiết dành cho Lập trình
    viên, Tester và Admin hệ thống. Thiết kế tuân theo chuẩn RESTful.


    ---

    ###  Hướng dẫn dành cho Người Dùng Mới (Getting Started)


    Nếu bạn là người mới sử dụng API này, hãy làm theo các bước sau để xác thực:


    1. **Đăng nhập (Login):** Go to the `/api/auth/login` route và nhập Email &
    Password.

    2. **Lấy Token:** Copy chuỗi `token` trả về từ kết quả JSON (trong object
    `data`).

    3. **Cấp quyền (Authorize):** Kéo lên trên cùng của trang web này, bấm vào
    nút **Authorize** màu xanh lá cây hoặc click vào biểu tượng 🔒 ở bất kỳ API
    nào, dán chuỗi token vừa copy vào ô **Value** và nhấn **Authorize**.

    4. **Bắt đầu gọi API:** Nhấn `Try it out` ở các endpoint có yêu cầu xác thực
    để thực hiện các yêu cầu (Requests). Server sẽ nhận diện được phiên làm việc
    của bạn.


    > ** Mẹo:** Swagger UI đã được cấu hình lưu lại Token kể cả khi bạn tải lại
    trang (Persist Authorization).


    ---


    ###  Mã Lỗi (Error Codes)

    Junkio API sử dụng hệ thống `Error Codes` chuẩn và thông báo lỗi đa ngôn ngữ
    (i18n). Thay vì đọc chuỗi ký tự thô, Frontend sẽ nhận được các mã như
    `WALLET_NOT_FOUND`, `INSUFFICIENT_BALANCE` và tự thông dịch thành văn bản.
...
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /api/admin/dashboard:
    get:
      summary: Lấy tổng quan hệ thống (Platform Dashboard)
      description: |
        Trả về các chỉ số tổng quan nhanh của toàn hệ thống bao gồm:
        - Tổng số người dùng đăng ký
        - Tổng số giao dịch đã ghi nhận
        - Tổng số nhóm gia đình đang hoạt động
        - Danh sách 5 người dùng đăng ký gần nhất

        API này được gọi khi Admin mở trang Dashboard.
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dữ liệu dashboard admin thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                      totalTransactions:
                        type: integer
                      totalFamilies:
                        type: integer
                      recentUsers:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            email:
                              type: string
                            role:
                              type: string
                              enum:
                                - member
                                - staff
                                - admin
                            createdAt:
                              type: string
                              format: date-time
              example:
                status: success
                message: Lấy dữ liệu Dashboard thành công
                data:
                  totalUsers: 1450
                  totalTransactions: 250000
                  totalFamilies: 35
                  recentUsers:
                    - id: b2df0d5d-1234-4abc-9def-bbbd02910001
                      name: Nguyễn Văn A
                      email: nguyenvana@junkio.com
                      role: member
                      createdAt: '2026-03-30T10:00:00.000Z'
                    - id: c3ef1e6e-5678-4bcd-aef0-ccce13a20002
                      name: Trần Thị B
                      email: tranthib@junkio.com
                      role: member
                      createdAt: '2026-03-29T08:30:00.000Z'
        '401':
          description: Chưa đăng nhập hoặc token hết hạn
        '403':
          description: Không có quyền admin
  /api/admin/analytics:
    get:
      summary: Lấy phân tích dữ liệu toàn hệ thống (System Analytics)
      description: >
        Trả về dữ liệu phân tích chuyên sâu bao gồm:

        - **Thống kê tổng hợp**: Tổng số ví, mục tiêu, ngân sách trên toàn hệ
        thống

        - **Biểu đồ tăng trưởng người dùng** theo từng tháng

        - **Top danh mục chi tiêu** phổ biến nhất

        - **Hoạt động hàng tuần** (số giao dịch mỗi ngày trong tuần gần nhất)


        Dùng để hiển thị biểu đồ trên trang Analytics của Admin Panel.
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dữ liệu analytics thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      stats:
                        type: object
                        properties:
                          totalWallets:
                            type: integer
                          totalGoals:
                            type: integer
                          totalBudgets:
                            type: integer
                      userGrowth:
                        type: array
                        items:
                          type: object
                          properties:
                            month:
                              type: string
                            count:
                              type: integer
                      topCategories:
                        type: array
                        items:
                          type: object
                      weeklyActivity:
                        type: array
                        items:
                          type: object
              example:
                status: success
                message: Lấy analytics thành công
                data:
                  stats:
                    totalWallets: 3200
                    totalGoals: 580
                    totalBudgets: 420
                  userGrowth:
                    - month: 2026-01
                      count: 120
                    - month: 2026-02
                      count: 185
                    - month: 2026-03
                      count: 210
                  topCategories:
                    - name: Ăn uống
                      count: 45000
                    - name: Di chuyển
                      count: 28000
                  weeklyActivity:
                    - day: Mon
                      transactions: 1200
                    - day: Tue
                      transactions: 1350
        '401':
          description: Chưa đăng nhập hoặc token hết hạn
        '403':
          description: Không có quyền admin
  /api/admin/financial-overview:
    get:
      summary: Lấy tổng quan tài chính toàn hệ thống (Financial Overview)
      description: >
        Cung cấp cái nhìn toàn cảnh về tình hình tài chính trên nền tảng:

        - **Tổng số dư hệ thống** (tổng balance của tất cả ví)

        - **Xu hướng thu/chi** theo từng tháng (dùng để vẽ biểu đồ line chart)

        - **Top người chi tiêu nhiều nhất**

        - **Tỷ lệ tuân thủ ngân sách** (% user giữ chi tiêu trong giới hạn
        budget)


        Endpoint này hữu ích để Admin đánh giá sức khỏe tài chính tổng thể.
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tổng quan tài chính thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      systemBalance:
                        type: number
                      revenueTrends:
                        type: array
                        items:
                          type: object
                          properties:
                            month:
                              type: string
                            income:
                              type: number
                            expense:
                              type: number
                      topSpenders:
                        type: array
                        items:
                          type: object
                      budgetCompliance:
                        type: number
              example:
                status: success
                message: Lấy tổng quan tài chính thành công
                data:
                  systemBalance: 15750000000
                  revenueTrends:
                    - month: 2026-01
                      income: 850000000
                      expense: 620000000
                    - month: 2026-02
                      income: 920000000
                      expense: 710000000
                  topSpenders:
                    - name: Nguyễn Văn C
                      totalExpense: 25000000
                    - name: Lê Thị D
                      totalExpense: 18500000
                  budgetCompliance: 72.5
        '401':
          description: Chưa đăng nhập hoặc token hết hạn
        '403':
          description: Không có quyền admin
...
  /api/admin/users/{id}/toggle-lock:
    put:
      summary: Khóa hoặc mở khóa tài khoản người dùng
      description: >
        Chuyển đổi trạng thái khóa của một tài khoản:

        - Nếu user đang **active** -> chuyển sang **locked** (không thể đăng
        nhập)

        - Nếu user đang **locked** -> chuyển sang **active** (cho phép đăng nhập
        lại)


        **Lưu ý:** Admin không thể khóa chính mình.

        Khi user bị khóa, token hiện tại vẫn hoạt động cho đến khi hết hạn.
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID của user cần khóa/mở khóa
      responses:
        '200':
          description: Thay đổi trạng thái khóa thành công
          content:
            application/json:
              example:
                status: success
                message: Đã khóa tài khoản user
                data:
                  id: b2c3d4e5-f6a7-8901-bcde-f12345678901
                  is_locked: true
        '400':
          description: Không được khóa chính mình (CANNOT_LOCK_SELF)
  /api/admin/users/{id}/role:
    put:
      summary: Thay đổi vai trò (role) của người dùng
      description: |
        Nâng cấp hoặc hạ cấp quyền hạn của một user trong hệ thống.
        Các vai trò hợp lệ:
        - `member`: Người dùng thường (mặc định khi đăng ký)
        - `staff`: Nhân viên hỗ trợ (có quyền xem báo cáo)
        - `admin`: Quản trị viên toàn quyền

        **Lưu ý:** Admin không thể thay đổi role của chính mình.
      tags:
        - Admin
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID của user cần đổi role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum:
                    - member
                    - staff
                    - admin
                  example: staff
            example:
              role: staff
      responses:
        '200':
          description: Đổi role thành công
          content:
            application/json:
              example:
                status: success
                message: Đổi role thành công
                data:
                  id: b2c3d4e5-f6a7-8901-bcde-f12345678901
                  role: staff
        '400':
          description: >-
            Role không hợp lệ hoặc đang tự đổi role (INVALID_ROLE /
            CANNOT_CHANGE_OWN_ROLE)
...